[MID-80] restructure assets folder
