+++ jsconfig.json
@@ -3,6 +3,7 @@
     "paths": {
       "@/*": ["./src/*"],
       "@icons/*": ["./public/assets/icons/*"],
+      "@images/*": ["./public/assets/images/*"],
       "@components": ["./src/app/_components/index"],
       "@hooks": ["./src/app/_hooks/index"],
       "@utils/*": ["./src/app/utils/*"],
